fix(predict): share pulsing live dot
